Character animation and movement

+++ src/by/segg3r/slicktest/math/Point.java
@@ -2,28 +2,28 @@ package by.segg3r.slicktest.math;
 
 public class Point {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
